// services/http-auth/src/generated/grpc/typescript/threads_pb.d.ts
export namespace PingPongMessage {
    export type AsObject = {
        name?: google_protobuf_wrappers_pb.StringValue.AsObject,
        timestamp?: google_protobuf_wrappers_pb.Int64Value.AsObject,
    }
}

export class ThreadModel extends jspb.Message { 

    hasId(): boolean;
    clearId(): void;
    getId(): google_protobuf_wrappers_pb.StringValue | undefined;
    setId(value?: google_protobuf_wrappers_pb.StringValue): ThreadModel;

    hasName(): boolean;
    clearName(): void;
    getName(): google_protobuf_wrappers_pb.StringValue | undefined;
    setName(value?: google_protobuf_wrappers_pb.StringValue): ThreadModel;

    hasDescription(): boolean;
    clearDescription(): void;
    getDescription(): google_protobuf_wrappers_pb.StringValue | undefined;
    setDescription(value?: google_protobuf_wrappers_pb.StringValue): ThreadModel;

    hasContent(): boolean;
    clearContent(): void;
    getContent(): google_protobuf_wrappers_pb.StringValue | undefined;
    setContent(value?: google_protobuf_wrappers_pb.StringValue): ThreadModel;
    clearImagesList(): void;
    getImagesList(): Array<google_protobuf_wrappers_pb.StringValue>;
    setImagesList(value: Array<google_protobuf_wrappers_pb.StringValue>): ThreadModel;
    addImages(value?: google_protobuf_wrappers_pb.StringValue, index?: number): google_protobuf_wrappers_pb.StringValue;

    hasCreatedAt(): boolean;
    clearCreatedAt(): void;
    getCreatedAt(): google_protobuf_wrappers_pb.StringValue | undefined;
    setCreatedAt(value?: google_protobuf_wrappers_pb.StringValue): ThreadModel;

    hasUpdatedAt(): boolean;
    clearUpdatedAt(): void;
    getUpdatedAt(): google_protobuf_wrappers_pb.StringValue | undefined;
    setUpdatedAt(value?: google_protobuf_wrappers_pb.StringValue): ThreadModel;

    hasOwner(): boolean;
    clearOwner(): void;
    getOwner(): google_protobuf_wrappers_pb.StringValue | undefined;
    setOwner(value?: google_protobuf_wrappers_pb.StringValue): ThreadModel;

    hasDeleted(): boolean;
    clearDeleted(): void;
    getDeleted(): google_protobuf_wrappers_pb.BoolValue | undefined;
    setDeleted(value?: google_protobuf_wrappers_pb.BoolValue): ThreadModel;
    clearPromotedList(): void;
    getPromotedList(): Array<PromotedObject>;
    setPromotedList(value: Array<PromotedObject>): ThreadModel;
    addPromoted(value?: PromotedObject, index?: number): PromotedObject;
    clearDonationsList(): void;
    getDonationsList(): Array<DonationObject>;
    setDonationsList(value: Array<DonationObject>): ThreadModel;
    addDonations(value?: DonationObject, index?: number): DonationObject;
    clearLikesList(): void;
    getLikesList(): Array<google_protobuf_wrappers_pb.StringValue>;
    setLikesList(value: Array<google_protobuf_wrappers_pb.StringValue>): ThreadModel;
    addLikes(value?: google_protobuf_wrappers_pb.StringValue, index?: number): google_protobuf_wrappers_pb.StringValue;
    clearDislikesList(): void;
    getDislikesList(): Array<google_protobuf_wrappers_pb.StringValue>;
    setDislikesList(value: Array<google_protobuf_wrappers_pb.StringValue>): ThreadModel;
    addDislikes(value?: google_protobuf_wrappers_pb.StringValue, index?: number): google_protobuf_wrappers_pb.StringValue;
    clearTagsList(): void;
    getTagsList(): Array<google_protobuf_wrappers_pb.StringValue>;
    setTagsList(value: Array<google_protobuf_wrappers_pb.StringValue>): ThreadModel;
    addTags(value?: google_protobuf_wrappers_pb.StringValue, index?: number): google_protobuf_wrappers_pb.StringValue;

    hasLikesCount(): boolean;
    clearLikesCount(): void;
    getLikesCount(): google_protobuf_wrappers_pb.UInt64Value | undefined;
    setLikesCount(value?: google_protobuf_wrappers_pb.UInt64Value): ThreadModel;

    hasDislikesCount(): boolean;
    clearDislikesCount(): void;
    getDislikesCount(): google_protobuf_wrappers_pb.UInt64Value | undefined;
    setDislikesCount(value?: google_protobuf_wrappers_pb.UInt64Value): ThreadModel;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): ThreadModel.AsObject;
    static toObject(includeInstance: boolean, msg: ThreadModel): ThreadModel.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: ThreadModel, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): ThreadModel;
    static deserializeBinaryFromReader(message: ThreadModel, reader: jspb.BinaryReader): ThreadModel;
}

export namespace ThreadModel {
    export type AsObject = {
        id?: google_protobuf_wrappers_pb.StringValue.AsObject,
        name?: google_protobuf_wrappers_pb.StringValue.AsObject,
        description?: google_protobuf_wrappers_pb.StringValue.AsObject,
        content?: google_protobuf_wrappers_pb.StringValue.AsObject,
        imagesList: Array<google_protobuf_wrappers_pb.StringValue.AsObject>,
        createdAt?: google_protobuf_wrappers_pb.StringValue.AsObject,
        updatedAt?: google_protobuf_wrappers_pb.StringValue.AsObject,
        owner?: google_protobuf_wrappers_pb.StringValue.AsObject,
        deleted?: google_protobuf_wrappers_pb.BoolValue.AsObject,
        promotedList: Array<PromotedObject.AsObject>,
        donationsList: Array<DonationObject.AsObject>,
        likesList: Array<google_protobuf_wrappers_pb.StringValue.AsObject>,
        dislikesList: Array<google_protobuf_wrappers_pb.StringValue.AsObject>,
        tagsList: Array<google_protobuf_wrappers_pb.StringValue.AsObject>,
        likesCount?: google_protobuf_wrappers_pb.UInt64Value.AsObject,
        dislikesCount?: google_protobuf_wrappers_pb.UInt64Value.AsObject,
    }
}
// ...
export class CreateRequest extends jspb.Message { 

    hasName(): boolean;
    clearName(): void;
    getName(): google_protobuf_wrappers_pb.StringValue | undefined;
    setName(value?: google_protobuf_wrappers_pb.StringValue): CreateRequest;

    hasDescription(): boolean;
    clearDescription(): void;
    getDescription(): google_protobuf_wrappers_pb.StringValue | undefined;
    setDescription(value?: google_protobuf_wrappers_pb.StringValue): CreateRequest;

    hasContent(): boolean;
    clearContent(): void;
    getContent(): google_protobuf_wrappers_pb.StringValue | undefined;
    setContent(value?: google_protobuf_wrappers_pb.StringValue): CreateRequest;

    hasOwner(): boolean;
    clearOwner(): void;
    getOwner(): google_protobuf_wrappers_pb.StringValue | undefined;
    setOwner(value?: google_protobuf_wrappers_pb.StringValue): CreateRequest;
    clearImagesList(): void;
    getImagesList(): Array<google_protobuf_wrappers_pb.StringValue>;
    setImagesList(value: Array<google_protobuf_wrappers_pb.StringValue>): CreateRequest;
    addImages(value?: google_protobuf_wrappers_pb.StringValue, index?: number): google_protobuf_wrappers_pb.StringValue;
    clearTagsList(): void;
    getTagsList(): Array<google_protobuf_wrappers_pb.StringValue>;
    setTagsList(value: Array<google_protobuf_wrappers_pb.StringValue>): CreateRequest;
    addTags(value?: google_protobuf_wrappers_pb.StringValue, index?: number): google_protobuf_wrappers_pb.StringValue;

    hasIspromoted(): boolean;
    clearIspromoted(): void;
    getIspromoted(): google_protobuf_wrappers_pb.BoolValue | undefined;
    setIspromoted(value?: google_protobuf_wrappers_pb.BoolValue): CreateRequest;

    hasAuth(): boolean;
    clearAuth(): void;
    getAuth(): google_protobuf_wrappers_pb.StringValue | undefined;
    setAuth(value?: google_protobuf_wrappers_pb.StringValue): CreateRequest;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): CreateRequest.AsObject;
    static toObject(includeInstance: boolean, msg: CreateRequest): CreateRequest.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: CreateRequest, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): CreateRequest;
    static deserializeBinaryFromReader(message: CreateRequest, reader: jspb.BinaryReader): CreateRequest;
}

export namespace CreateRequest {
    export type AsObject = {
        name?: google_protobuf_wrappers_pb.StringValue.AsObject,
        description?: google_protobuf_wrappers_pb.StringValue.AsObject,
        content?: google_protobuf_wrappers_pb.StringValue.AsObject,
        owner?: google_protobuf_wrappers_pb.StringValue.AsObject,
        imagesList: Array<google_protobuf_wrappers_pb.StringValue.AsObject>,
        tagsList: Array<google_protobuf_wrappers_pb.StringValue.AsObject>,
        ispromoted?: google_protobuf_wrappers_pb.BoolValue.AsObject,
        auth?: google_protobuf_wrappers_pb.StringValue.AsObject,
    }
}
// ...
export namespace AmountWithAuthRequest {
    export type AsObject = {
        auth?: google_protobuf_wrappers_pb.StringValue.AsObject,
        amount?: google_protobuf_wrappers_pb.UInt64Value.AsObject,
        threadId?: google_protobuf_wrappers_pb.StringValue.AsObject,
    }
}

export class Pagination extends jspb.Message { 

    hasPage(): boolean;
    clearPage(): void;
    getPage(): google_protobuf_wrappers_pb.UInt64Value | undefined;
    setPage(value?: google_protobuf_wrappers_pb.UInt64Value): Pagination;

    hasLimit(): boolean;
    clearLimit(): void;
    getLimit(): google_protobuf_wrappers_pb.UInt64Value | undefined;
    setLimit(value?: google_protobuf_wrappers_pb.UInt64Value): Pagination;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): Pagination.AsObject;
    static toObject(includeInstance: boolean, msg: Pagination): Pagination.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: Pagination, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): Pagination;
    static deserializeBinaryFromReader(message: Pagination, reader: jspb.BinaryReader): Pagination;
}

export namespace Pagination {
    export type AsObject = {
        page?: google_protobuf_wrappers_pb.UInt64Value.AsObject,
        limit?: google_protobuf_wrappers_pb.UInt64Value.AsObject,
    }
}
// ...
export namespace WalletWithAuthRequest {
    export type AsObject = {
        wallet?: google_protobuf_wrappers_pb.StringValue.AsObject,
        id?: google_protobuf_wrappers_pb.StringValue.AsObject,
    }
}

export class IdRequest extends jspb.Message { 

    hasId(): boolean;
    clearId(): void;
    getId(): google_protobuf_wrappers_pb.StringValue | undefined;
    setId(value?: google_protobuf_wrappers_pb.StringValue): IdRequest;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): IdRequest.AsObject;
    static toObject(includeInstance: boolean, msg: IdRequest): IdRequest.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: IdRequest, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): IdRequest;
    static deserializeBinaryFromReader(message: IdRequest, reader: jspb.BinaryReader): IdRequest;
}

export namespace IdRequest {
    export type AsObject = {
        id?: google_protobuf_wrappers_pb.StringValue.AsObject,
    }
}
// ...
export namespace StatsModel {
    export type AsObject = {
        id?: google_protobuf_wrappers_pb.StringValue.AsObject,
        name?: google_protobuf_wrappers_pb.StringValue.AsObject,
        promoted?: PromotedStats.AsObject,
        donations?: DonationsStats.AsObject,
        likesCount?: google_protobuf_wrappers_pb.Int64Value.AsObject,
        dislikesCount?: google_protobuf_wrappers_pb.Int64Value.AsObject,
    }
}