// services/http-auth/src/generated/grpc/typescript/threads_grpc_pb.d.ts
export interface IThreadsServer extends grpc.UntypedServiceImplementation {
    pingPong: grpc.handleUnaryCall<threads_pb.PingPongMessage, threads_pb.PingPongMessage>;
    create: grpc.handleUnaryCall<threads_pb.CreateRequest, threads_pb.ThreadModel>;
    update: grpc.handleUnaryCall<threads_pb.ThreadModel, threads_pb.ThreadModel>;
    delete: grpc.handleUnaryCall<threads_pb.IdRequest, threads_pb.ThreadModel>;
    getMany: grpc.handleServerStreamingCall<threads_pb.Pagination, threads_pb.ThreadModel>;
    getTotalCount: grpc.handleUnaryCall<google_protobuf_empty_pb.Empty, google_protobuf_wrappers_pb.Int32Value>;
    getByOwner: grpc.handleServerStreamingCall<threads_pb.IdRequest, threads_pb.ThreadModel>;
    getOne: grpc.handleUnaryCall<threads_pb.IdRequest, threads_pb.ThreadModel>;
    getStatistics: grpc.handleServerStreamingCall<threads_pb.Pagination, threads_pb.StatsModel>;
    getLikes: grpc.handleServerStreamingCall<threads_pb.IdRequest, google_protobuf_wrappers_pb.StringValue>;
    getDislikes: grpc.handleServerStreamingCall<threads_pb.IdRequest, google_protobuf_wrappers_pb.StringValue>;
    like: grpc.handleUnaryCall<threads_pb.WalletWithAuthRequest, google_protobuf_empty_pb.Empty>;
    dislike: grpc.handleUnaryCall<threads_pb.WalletWithAuthRequest, google_protobuf_empty_pb.Empty>;
    donate: grpc.handleUnaryCall<threads_pb.AmountWithAuthRequest, google_protobuf_empty_pb.Empty>;
    promote: grpc.handleUnaryCall<threads_pb.AmountWithAuthRequest, google_protobuf_empty_pb.Empty>;
}